requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==4.1.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
flake8==6.0.0
inflection==0.5.1
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2022.6
